Сделал единую иерархию папок в user_service. Прогнал все тесты.
