+++ src/main/java/school/faang/user_service/dto/goal/GoalDto.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.dto;
+package school.faang.user_service.dto.goal;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import school.faang.user_service.entity.goal.GoalStatus;
 
 import java.util.List;
+
 @Data
 @Builder
 @NoArgsConstructor
