+++ src/main/java/school/faang/user_service/exception/DataValidationException.java
@@ -2,6 +2,6 @@ package school.faang.user_service.exception;
 
 public class DataValidationException extends RuntimeException {
     public DataValidationException(String exception) {
-       super(exception);
+        super(exception);
     }
 }
