+++ src/main/java/school/faang/user_service/filter/goal/GoalDescriptionFilter.java
@@ -1,8 +1,8 @@
 package school.faang.user_service.filter.goal;
 
 import org.springframework.stereotype.Component;
-import school.faang.user_service.dto.GoalDto;
-import school.faang.user_service.dto.GoalFilterDto;
+import school.faang.user_service.dto.goal.GoalDto;
+import school.faang.user_service.dto.goal.GoalFilterDto;
 
 import java.util.List;
 import java.util.Objects;
