+++ src/main/java/school/faang/user_service/filter/user/UserExperienceFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.user.filter;
+package school.faang.user_service.filter.user;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.user.UserFilterDto;
@@ -7,14 +7,14 @@ import school.faang.user_service.entity.User;
 import java.util.stream.Stream;
 
 @Component
-public class UserCityFilter implements UserFilter {
+public class UserExperienceFilter implements UserFilter {
     @Override
     public boolean isApplicable(UserFilterDto filters) {
-        return filters.getCity() != null;
+        return filters.getExperience() != null;
     }
 
     @Override
     public Stream<User> apply(Stream<User> users, UserFilterDto filters) {
-        return users.filter(user -> user.getCity().contains(filters.getCity()));
+        return users.filter(user -> user.getExperience().equals(filters.getExperience()));
     }
 }
