+++ src/main/java/school/faang/user_service/filter/user/UserFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.user.filter;
+package school.faang.user_service.filter.user;
 
 import school.faang.user_service.dto.user.UserFilterDto;
 import school.faang.user_service.entity.User;
