+++ src/main/java/school/faang/user_service/mapper/GoalMapper.java
@@ -1,11 +1,7 @@
 package school.faang.user_service.mapper;
 
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.MappingTarget;
-import org.mapstruct.Named;
-import org.mapstruct.ReportingPolicy;
-import school.faang.user_service.dto.GoalDto;
+import org.mapstruct.*;
+import school.faang.user_service.dto.goal.GoalDto;
 import school.faang.user_service.entity.Skill;
 import school.faang.user_service.entity.goal.Goal;
