+++ src/main/java/school/faang/user_service/mapper/SkillMapper.java
@@ -1,8 +1,7 @@
-package school.faang.user_service.mappers;
+package school.faang.user_service.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.skill.SkillCandidateDto;
 import school.faang.user_service.dto.skill.SkillDto;
@@ -12,7 +11,9 @@ import school.faang.user_service.entity.Skill;
 @Mapper(componentModel = "spring")
 public interface SkillMapper {
     Skill toEntity(SkillDto skillDto);
+
     SkillDto toDto(Skill skill);
+
     @Mapping(target = "skillDto.id", source = "skill.id")
     @Mapping(target = "skillDto.title", source = "skill.title")
     SkillCandidateDto skillToSkillCandidateDto(Skill skill, long offersAmount);
