+++ src/main/java/school/faang/user_service/validator/SkillValidator.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.validation;
+package school.faang.user_service.validator;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
