+++ src/test/java/school/faang/user_service/exception/DataValidationExceptionTest.java
@@ -1,9 +1,8 @@
-package school.faang.user_service.service.exception;
+package school.faang.user_service.exception;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
-import school.faang.user_service.exception.DataValidationException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
